use executable, but connect through TCP
Using chicken/guile-lsp-server through stdio didn't work quite well. The problem
is that any messages written by Guile to stderr (like "warnings" when loading
the LSP server) confused VS Code and blocked communication with the LSP
Server. One could investigate a way to redirect the executable's stderr, but
another approach is to simply rely on TCP to do the communication, as this
commit does.