ignore server's stdio
The client was crashing when trying to autocomplete (or resolve) an identifier.
I think somehow the LSP server's stderr was interfering with the LSP client
(I don't know the internals of VSCode's LSP client library). By ignoring the
LSP server's stdio the problem is gone. I still have to figure out how to print
debug messages when needed though.