Disable auto-install of chicken-lsp-server
Auto-install of chicken-lsp-server involved installing it locally to the
extension. This can be extremely slow though, specially on fresh installs
of chicken, since all dependencies are installed in a custom directory. I didn't
find a way to run a command with sudo privilege from inside vscodium (didn't
try much either, any ideas are welcome).
Since installing the LSP server is pretty easy, we just tell the user to run
`chicken-install -s lsp-server` and reopen the window.